# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "Advertiser": {
        "properties": {
          "data_source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "leads_file": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "review_available": {
            "description": "True once both files are uploaded",
            "title": "Review Available",
            "type": "boolean"
          },
          "stage_history_file": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileProfile"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "data_source",
          "leads_file",
          "stage_history_file",
          "review_available"
        ],
        "title": "Advertiser",
        "type": "object"
      },
      "Column": {
        "properties": {
          "examples": {
            "description": "The column's first few distinct non-empty values",
            "items": {
              "type": "string"
            },
            "title": "Examples",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "examples"
        ],
        "title": "Column",
        "type": "object"
      },
      "ColumnKind": {
        "description": "How an input column is read: as a number, or as one of a set of categories.",
        "enum": [
          "number",
          "category"
        ],
        "title": "ColumnKind",
        "type": "string"
      },
      "CrmStage": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "row_count"
        ],
        "title": "CrmStage",
        "type": "object"
      },
      "DataSource": {
        "description": "Where an advertiser's data comes from; every result is labelled with it.",
        "enum": [
          "hand_made_test",
          "simulated",
          "public",
          "private"
        ],
        "title": "DataSource",
        "type": "string"
      },
      "FileKind": {
        "enum": [
          "leads",
          "stage-history"
        ],
        "title": "FileKind",
        "type": "string"
      },
      "FileProfile": {
        "properties": {
          "column_names": {
            "items": {
              "type": "string"
            },
            "title": "Column Names",
            "type": "array"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/FileKind"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "file_name",
          "uploaded_at",
          "row_count",
          "column_names"
        ],
        "title": "FileProfile",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Health": {
        "properties": {
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "Health",
        "type": "object"
      },
      "LadderPlace": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          }
        },
        "required": [
          "place",
          "name"
        ],
        "title": "LadderPlace",
        "type": "object"
      },
      "LeadsColumns": {
        "description": "Which leads-file column holds what. Every column not marked here is dropped.",
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The lead's email, to be scrambled",
            "title": "Email"
          },
          "inputs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ColumnKind"
            },
            "description": "The inputs to the score, each read as its kind",
            "title": "Inputs",
            "type": "object"
          },
          "lead_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The lead's name, to be removed",
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The lead's phone, to be scrambled",
            "title": "Phone"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the lead was submitted",
            "title": "Submitted At"
          }
        },
        "title": "LeadsColumns",
        "type": "object"
      },
      "Lost": {
        "enum": [
          "lost"
        ],
        "title": "Lost",
        "type": "string"
      },
      "Mapping": {
        "properties": {
          "crm_stages": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Place"
            },
            "description": "Where each CRM stage name sits: a Stage, or Lost",
            "title": "Crm Stages",
            "type": "object"
          },
          "leads": {
            "$ref": "#/components/schemas/LeadsColumns",
            "default": {
              "inputs": {}
            }
          },
          "stage_history": {
            "$ref": "#/components/schemas/StageHistoryColumns",
            "default": {}
          },
          "typical_deal_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typical Deal Size"
          }
        },
        "title": "Mapping",
        "type": "object"
      },
      "MappingReview": {
        "properties": {
          "confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When a person confirmed it; null in draft",
            "title": "Confirmed At"
          },
          "crm_stages": {
            "description": "Every CRM stage name in the column marked as the CRM stage, most used first",
            "items": {
              "$ref": "#/components/schemas/CrmStage"
            },
            "title": "Crm Stages",
            "type": "array"
          },
          "mapping": {
            "$ref": "#/components/schemas/Mapping"
          },
          "places": {
            "description": "Where a CRM stage can be placed: the Canonical ladder in order, then Lost",
            "items": {
              "$ref": "#/components/schemas/LadderPlace"
            },
            "title": "Places",
            "type": "array"
          },
          "problems": {
            "description": "Why it cannot be confirmed yet; empty once it can",
            "items": {
              "type": "string"
            },
            "title": "Problems",
            "type": "array"
          }
        },
        "required": [
          "mapping",
          "confirmed_at",
          "problems",
          "crm_stages",
          "places"
        ],
        "title": "MappingReview",
        "type": "object"
      },
      "NewAdvertiser": {
        "properties": {
          "data_source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "data_source"
        ],
        "title": "NewAdvertiser",
        "type": "object"
      },
      "Place": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/Stage"
          },
          {
            "$ref": "#/components/schemas/Lost"
          }
        ]
      },
      "Problem": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Stage": {
        "description": "A Stage of the Canonical ladder; members are in ladder order.",
        "enum": [
          "submitted",
          "contact_attempted",
          "engaged",
          "qualified",
          "proposal",
          "won"
        ],
        "title": "Stage",
        "type": "string"
      },
      "StageHistoryColumns": {
        "properties": {
          "changed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the change of CRM stage happened",
            "title": "Changed At"
          },
          "crm_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crm Stage"
          },
          "deal_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deal Value"
          },
          "lead_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Id"
          }
        },
        "title": "StageHistoryColumns",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Emva API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/advertisers": {
      "post": {
        "operationId": "createAdvertiser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewAdvertiser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Advertiser"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Advertiser"
      }
    },
    "/advertisers/{advertiser_id}": {
      "get": {
        "operationId": "getAdvertiser",
        "parameters": [
          {
            "in": "path",
            "name": "advertiser_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advertiser Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Advertiser"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Advertiser"
      }
    },
    "/advertisers/{advertiser_id}/files/{kind}": {
      "put": {
        "description": "Upload the file, replacing any earlier upload of the same kind.",
        "operationId": "uploadFile",
        "parameters": [
          {
            "in": "path",
            "name": "advertiser_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advertiser Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FileKind"
            }
          },
          {
            "in": "query",
            "name": "file_name",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "File Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/csv": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not a readable CSV file"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Another upload won, or the mapping is confirmed"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Over 20 MB"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not saved"
          }
        },
        "summary": "Upload File"
      }
    },
    "/advertisers/{advertiser_id}/files/{kind}/columns": {
      "get": {
        "description": "Each column of the file with its first few values, read from the file itself.",
        "operationId": "getColumns",
        "parameters": [
          {
            "in": "path",
            "name": "advertiser_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advertiser Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FileKind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Column"
                  },
                  "title": "Response Getcolumns",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Stored file unreadable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Columns"
      }
    },
    "/advertisers/{advertiser_id}/mapping": {
      "get": {
        "description": "The mapping as last saved (an empty draft at first), with what stops it being confirmed.",
        "operationId": "getMapping",
        "parameters": [
          {
            "in": "path",
            "name": "advertiser_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advertiser Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MappingReview"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Stored file unreadable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mapping"
      },
      "put": {
        "description": "Keep the draft as it stands; refused once the mapping is confirmed.",
        "operationId": "saveMapping",
        "parameters": [
          {
            "in": "path",
            "name": "advertiser_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advertiser Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mapping"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MappingReview"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Already confirmed, or a stored file is unreadable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Mapping"
      }
    },
    "/advertisers/{advertiser_id}/mapping/confirmation": {
      "post": {
        "description": "Confirm the saved draft, recording when; refused while it has problems.",
        "operationId": "confirmMapping",
        "parameters": [
          {
            "in": "path",
            "name": "advertiser_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advertiser Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MappingReview"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not confirmable yet, already confirmed, or a stored file unreadable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Mapping"
      }
    },
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}